(#83) Factor out KGBotka.Eval
