+++ kgbotka.cabal
@@ -77,6 +77,7 @@ executable kgbotka
                      , KGBotka.Bttv
                      , KGBotka.Http
                      , KGBotka.Ffz
+                     , KGBotka.Eval
                      , Control.Monad.Trans.Extra
                      , Control.Monad.Trans.Eval
